integer and float constants

+++ ast.c
@@ -31,7 +31,7 @@ Expr *expr_alloc(ExprKind kind, SourcePos pos) {
     return expr;
 }
 
-Expr *expr_int(SourcePos pos, int32_t val) {
+Expr *expr_int(SourcePos pos, int64_t val) {
     Expr *expr = expr_alloc(EXPR_INT, pos);
     expr->int_val = val;
     return expr;
